+++ flexmeasures/api/v3_0/__init__.py
@@ -40,9 +40,6 @@ from flexmeasures.api.v3_0.sources import SourceAPI
 from flexmeasures.api.v3_0.assets import (
     flex_context_schema_openAPI,
     AssetAPIQuerySchema,
-    DefaultAssetViewJSONSchema,
-    StatusPageTabJSONSchema,
-    StatusPageChildJobsJSONSchema,
 )
 from flexmeasures.data.schemas.annotations import AnnotationSchema
 from flexmeasures.data.schemas.automations import (
@@ -236,9 +233,6 @@ def create_openapi_specs(app: Flask):
         ("AutomationUpdateSchema", AutomationUpdateSchema),
         ("ReportTriggerSchema", ReportTriggerSchema),
         ("CopyAssetSchema", CopyAssetSchema),
-        ("DefaultAssetViewJSONSchema", DefaultAssetViewJSONSchema),
-        ("StatusPageTabJSONSchema", StatusPageTabJSONSchema),
-        ("StatusPageChildJobsJSONSchema", StatusPageChildJobsJSONSchema),
         ("AccountSchema", AccountSchema(partial=True)),
         ("AccountCreateSchema", AccountCreateSchema()),
         ("AccountPatchSchema", AccountPatchSchema()),